+++ src/components/DataDisplay/Avatar/stories/Avatar.stories.tsx
@@ -6,9 +6,9 @@ const avatarSrc = "https://avatars.githubusercontent.com/u/16801167?s=200&v=4";
 
 const TemplateImage: StoryFn<typeof Avatar> = (args) => (
   <Stack direction="row" spacing={2} alignItems="center" justifyContent="center" height="100%">
-    <Avatar src={avatarSrc} sx={{ height: 24, width: 24 }} {...args} />
-    <Avatar src={avatarSrc} {...args} />
-    <Avatar src={avatarSrc} sx={{ height: 56, width: 56 }} {...args} />
+    <Avatar src={args?.src || avatarSrc} sx={{ height: 24, width: 24 }} {...args} />
+    <Avatar src={args?.src || avatarSrc} {...args} />
+    <Avatar src={args?.src || avatarSrc} sx={{ height: 56, width: 56 }} {...args} />
   </Stack>
 );
 
@@ -50,6 +50,11 @@ const TemplateBadge: StoryFn<typeof Avatar> = (args) => (
 
 export const Image = TemplateImage.bind({});
 Image.args = {};
+
+export const WithoutImage = TemplateImage.bind({});
+WithoutImage.args = {
+  src: "null",
+};
 
 export const ImageSquare = TemplateImage.bind({});
 ImageSquare.args = {