+++ src/components/Inputs/ActionAppBar/index.ts
@@ -1,0 +1,4 @@
+import ActionAppBar from "./ActionAppBar";
+
+export default ActionAppBar;
+export * from "./ActionAppBar";