+++ src/components/Inputs/TextFieldAppBar/index.ts
@@ -1,0 +1,4 @@
+import TextFieldAppBar from "./TextFieldAppBar";
+
+export default TextFieldAppBar;
+export * from "./TextFieldAppBar";