+++ src/components/Navigation/NavigationMenu/SecondaryMenuButton/SecondaryMenuButton.tsx
@@ -2,6 +2,7 @@ import {
   alpha,
   Avatar,
   Backdrop,
+  Box,
   Button,
   IconButton,
   ListItemIcon,
@@ -36,14 +37,15 @@ const styles = {
   },
   button: {
     "&:hover": {
+      background: ({ palette }: Theme) => palette.grey[50],
       cursor: "pointer",
     },
-    borderRadius: 0,
-    borderTop: ({ palette }: Theme) => `solid 1px ${palette.mode === "dark" ? palette.divider : alpha(palette.common.white, 0.12)}`,
-    justifyContent: "space-between",
+    alignSelf: "flex-start",
+    borderRadius: ({ shape }: Theme) => `${shape.borderRadius}px`,
+    justifyContent: "flex-start",
     minHeight: "auto",
     paddingX: 3,
-    paddingY: 2,
+    paddingY: 1,
     width: "100%",
   },
   menuItem: {
@@ -77,26 +79,22 @@ const SecondaryMenuButton = ({ variant = "button", ...props }: SecondaryMenuButt
   const isIconButton = variant === "icon";
 
   return (
-    <>
+    <Box padding={isMobile ? 0 : 2} width="100%">
       {isButton && (
         <Button
           onClick={openMenu}
+          startIcon={secondaryMenu?.startIcon}
+          variant="text"
           sx={{
             ...styles.button,
             ...(isMobile && {
-              borderTop: "none",
               marginBottom: 1,
               paddingX: 3,
-              width: "auto",
             }),
-            color: ({ palette }: Theme) => {
-              const backgroundColor = palette.mode === "dark" ? palette.background.default : palette.primary.main;
-
-              return palette.getContrastText(backgroundColor);
-            },
+            color: "text.primary",
           }}
         >
-          <Stack spacing={2} alignItems="center" direction="row" component="span">
+          <Stack spacing={2} alignItems="center" direction="row" component="span" flex={1}>
             {secondaryMenu?.avatar && (
               <Avatar
                 src={secondaryMenu?.avatar?.src}
@@ -129,18 +127,12 @@ const SecondaryMenuButton = ({ variant = "button", ...props }: SecondaryMenuButt
             )}
           </Stack>
           {isButton &&
-            (secondaryMenu?.iconOpenMenu || (
-              <SvgIcon>
-                {isMobile ? (
+            (secondaryMenu?.iconOpenMenu ||
+              (isMobile && (
+                <SvgIcon>
                   <path d="M8.12 9.29 12 13.17l3.88-3.88c.39-.39 1.02-.39 1.41 0 .39.39.39 1.02 0 1.41l-4.59 4.59c-.39.39-1.02.39-1.41 0L6.7 10.7a.9959.9959 0 0 1 0-1.41c.39-.38 1.03-.39 1.42 0z" />
-                ) : (
-                  <path
-                    d="M12 8.25C13.1 8.25 14 7.35 14 6.25C14 5.15 13.1 4.25 12 4.25C10.9 4.25 10 5.15 10 6.25C10 7.35 10.9 8.25 12 8.25ZM12 10.25C10.9 10.25 10 11.15 10 12.25C10 13.35 10.9 14.25 12 14.25C13.1 14.25 14 13.35 14 12.25C14 11.15 13.1 10.25 12 10.25ZM12 16.25C10.9 16.25 10 17.15 10 18.25C10 19.35 10.9 20.25 12 20.25C13.1 20.25 14 19.35 14 18.25C14 17.15 13.1 16.25 12 16.25Z"
-                    fill="white"
-                  />
-                )}
-              </SvgIcon>
-            ))}
+                </SvgIcon>
+              )))}
         </Button>
       )}
 
@@ -204,7 +196,7 @@ const SecondaryMenuButton = ({ variant = "button", ...props }: SecondaryMenuButt
           })}
         </Menu>
       </Backdrop>
-    </>
+    </Box>
   );
 };