+++ src/components/Navigation/NavigationMenu/TextFieldSearch/index.ts
@@ -1,4 +1,0 @@
-import TextFieldSearch from "./TextFieldSearch";
-
-export default TextFieldSearch;
-export * from "./TextFieldSearch";