+++ src/config/theme.ts
@@ -110,6 +110,17 @@ const commonThemeOptions: ThemeOptions = {
         },
       ],
     },
+    MuiAvatar: {
+      styleOverrides: {
+        root: ({ theme }) => ({
+          "& svg": {
+            color: theme.palette.text.disabled,
+          },
+          backgroundColor: theme.palette.grey[100],
+          color: theme.palette.text.primary,
+        }),
+      },
+    },
     MuiBadge: {
       styleOverrides: {
         badge: ({ theme, ownerState }) => {
@@ -133,9 +144,31 @@ const commonThemeOptions: ThemeOptions = {
         },
       },
     },
+    MuiBottomNavigationAction: {
+      styleOverrides: {
+        label: ({ theme }) => ({
+          "&.Mui-selected": {
+            color: theme.palette.common.white,
+          },
+          color: theme.palette.text.secondary,
+        }),
+        root: ({ theme }) => ({
+          "&.Mui-selected": {
+            color: theme.palette.common.white,
+            svg: {
+              color: theme.palette.common.white,
+            },
+          },
+          backgroundColor: theme.palette.background.dark,
+        }),
+      },
+    },
     MuiButton: {
       styleOverrides: {
         root: ({ theme }) => ({
+          "&:not(:hover)": {
+            boxShadow: "none",
+          },
           borderRadius: theme.shape.borderRadius,
           lineHeight: 1,
           paddingLeft: "24px",
@@ -145,7 +178,10 @@ const commonThemeOptions: ThemeOptions = {
       },
       variants: [
         {
-          props: { color: "primary", variant: "outlined" },
+          props: {
+            color: "primary",
+            variant: "outlined",
+          },
           style: ({ theme }) => ({
             backgroundColor: alpha(theme.palette.primary.main, 0.08),
             borderColor: alpha(theme.palette.primary.main, 0.5),
@@ -170,10 +206,34 @@ const commonThemeOptions: ThemeOptions = {
           },
         },
         {
-          props: { variant: "contained" },
-          style: {
-            color: "white",
+          props: {
+            variant: "contained",
           },
+          style: ({ theme }) => ({
+            color: theme.palette.primary.contrastText,
+          }),
+        },
+        {
+          props: {
+            color: "primary",
+            variant: "contained",
+          },
+          style: ({ theme }) => ({
+            ...(theme.palette.mode === "dark" && { outline: `solid 1px ${theme.palette.divider}` }),
+            backgroundColor: theme.palette.mode === "light" ? theme.palette.primary.main : theme.palette.secondary.main,
+            color: theme.palette.mode === "light" ? theme.palette.primary.contrastText : theme.palette.secondary.contrastText,
+          }),
+        },
+        {
+          props: {
+            color: "secondary",
+            variant: "contained",
+          },
+          style: ({ theme }) => ({
+            ...(theme.palette.mode === "light" && { outline: `solid 1px ${theme.palette.divider}` }),
+            backgroundColor: theme.palette.mode === "light" ? theme.palette.secondary.main : theme.palette.primary.main,
+            color: theme.palette.mode === "light" ? theme.palette.secondary.contrastText : theme.palette.primary.contrastText,
+          }),
         },
         {
           props: { variant: "link" },
@@ -501,6 +561,9 @@ const commonThemeOptions: ThemeOptions = {
     MuiTab: {
       styleOverrides: {
         root: ({ theme }) => ({
+          "&.Mui-selected": {
+            color: theme.palette.text.primary,
+          },
           "&:first-of-type": {
             marginLeft: 0,
           },
@@ -520,6 +583,15 @@ const commonThemeOptions: ThemeOptions = {
         root: ({ theme }) => ({
           border: `solid 1px ${theme.palette.divider}`,
           borderRadius: theme.shape.borderRadius,
+        }),
+      },
+    },
+    MuiTabs: {
+      styleOverrides: {
+        root: ({ theme }) => ({
+          "& .MuiTabs-indicator": {
+            backgroundColor: theme.palette.text.primary,
+          },
         }),
       },
     },