+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.59.0
-- **[feat]** - add `Backoffice` component
-- **[feat]** - add `Header` component
-- **[feat]** - New variant `shape-background` for `Logo` component
+## v2.0.0-beta.1
+- **[feat]** - update theme colors
+- **[feat]** - add new`Backoffice` component
+- **[feat]** - add new`AppBar` component
+- **[feat]** - Add new `Logo`
+- **[feat]** - Add new `Logo` for supplier
+- **[feat]** - Add new `TextFieldAppBar` component
+- **[feat]** - Add new `ActionAppBar` component
+- **[feat]** - Add new `AvatarAppBar` component
+- **[feat]** - adapte `NavigationMenu` component for new `AppBar` component
+- **[feat]** - update `Button` theme
+- **[feat]** - update `Tabs` et `Tabs` theme
+- **[feat]** - improve `FeatureEnableProvider` initial feature
+- **[docs]** - Add new story for `AppBar` component
+- **[docs]** - Add new story for `Backoffice` component
 - **[chore]** - Improve module augmentation declaration